moved self service site analytics into its own tab
