+++ blocks/dashboard/dashboard.css
@@ -105,7 +105,7 @@
   -webkit-box-orient: vertical;
 }
 
-.dashboard.block :is(.sites, .admin) > p:only-child {
+.dashboard.block :is(.sites, .admin-section.users) > p:only-child {
   display: grid;
   place-items: center;
   min-height: calc(100vh - 350px);
@@ -165,23 +165,23 @@
   padding-bottom: 0;
 }
 
-dialog:has(.admin) {
+dialog:has(.admin-section.users) {
   width: auto;
 }
 
-:is(.dashboard.block, dialog) .admin .clusterize {
+:is(.dashboard.block, dialog) .admin-section.users .clusterize {
   margin-top: 48px;
   margin-bottom: 48px;
 }
 
-:is(.dashboard.block, dialog) .admin table {
+:is(.dashboard.block, dialog) .admin-section.users table {
   width: 100%;
   table-layout: fixed;
   overflow-x: auto;
   -webkit-overflow-scrolling: touch;
 }
 
-:is(.dashboard.block, dialog) .admin td.empty {
+:is(.dashboard.block, dialog) .admin-section.users td.empty {
   text-align: center;
   text-transform: uppercase;
   padding: 32px;
@@ -189,67 +189,67 @@ dialog:has(.admin) {
   font-size: 12px;
 }
 
-.dashboard.block .details .admin img,
-dialog .admin img {
+.dashboard.block .details .admin-section.users img,
+dialog .admin-section.users img {
   height: 64px;
   width: 64px;
 }
 
-.dashboard.block .admin .filter {
+.dashboard.block .admin-section.users .filter {
   margin-top: 16px;
 }
 
-:is(.dashboard.block, dialog) .admin table :is(th, td) {
+:is(.dashboard.block, dialog) .admin-section.users table :is(th, td) {
   max-width: 100%;
   overflow: auto;
   word-wrap: break-word;
 }
 
-:is(.dashboard.block) .admin table :is(th) {
+:is(.dashboard.block) .admin-section.users table :is(th) {
   position: sticky;
   top: 0;
   background-color: var(--white);
 }
 
-:is(dialog) .admin table :is(th) {
+:is(dialog) .admin-section.users table :is(th) {
   position: sticky;
   top: 0;
   background-color: var(--white);
 }
 
-:is(.dashboard.block, dialog) .admin table td:first-child {
+:is(.dashboard.block, dialog) .admin-section.users table td:first-child {
   font-weight: 500;
 }
 
-:is(.dashboard.block, dialog) .admin table [data-column="event"] {
+:is(.dashboard.block, dialog) .admin-section.users table [data-column="event"] {
   text-transform: capitalize;
 }
 
-.dashboard.block .admin table + .button-container:has(button:only-child) {
+.dashboard.block .admin-section.users table + .button-container:has(button:only-child) {
   display: none;
 }
 
-.dashboard.block .admin h2 {
+.dashboard.block .admin-section.users h2 {
   padding-block: 20px;
   font-size: 20px;
   line-height: 24px;
 }
 
-.dashboard.block .admin h3 {
+.dashboard.block .admin-section.users h3 {
   font-size: 16px;
   padding-bottom: 12px;
 }
 
-body:not(.is-headwire, .is-test-user) .dashboard.block .admin {
+body:not(.is-headwire, .is-test-user) .dashboard.block .admin-section.users {
   display: none;
 }
 
 @media (width < 768px) {
-  .dashboard.block .tab-content.admin.is-selected > div {
+  .dashboard.block .tab-content.admin-section.users.is-selected > div {
     overflow-x: auto;
   }
 
-  :is(.dashboard.block, dialog) .admin table {
+  :is(.dashboard.block, dialog) .admin-section.users table {
     table-layout: auto;
   }
